+++ data-raw/data.R
@@ -1,4 +1,95 @@
-# This file generates and writes the dataset 'mtcars_df' and 'mtcars_df2'
+# This file generates and writes demo datasets
+
+
+# demo_df and demo_a11ytable (as of v0.3) ---------------------------------
+
+
+set.seed(1066)
+
+cover_list <- list(
+  "Section 1" = c("First row of Section 1.", "Second row of Section 1."),
+  "Section 2" = "The only row of Section 2.",
+  "Section 3" = c(
+    "[Website](https://co-analysis.github.io/a11ytables/)",
+    "[Email address](mailto:fake.address@a11ytables.com)"
+  )
+)
+
+contents_df <- data.frame(
+  "Sheet name" = c("Notes", "Table 1", "Table 2"),
+  "Sheet title" = c(
+    "Notes used in this workbook",
+    "First Example Sheet",
+    "Second Example Sheet"
+  ),
+  check.names = FALSE
+)
+
+notes_df <- data.frame(
+  "Note number" = paste0("[note ", 1:3, "]"),
+  "Note text" = c("First note.", "Second note.", "Third note."),
+  check.names = FALSE
+)
+
+table_1_df <- data.frame(
+  Category = LETTERS[1:10],
+  "Numeric [note 1]" = 1:10,
+  "Numeric suppressed" = c(1:4, "[c]", 6:9, "[x]"),
+  "Numeric thousands" = abs(round(rnorm(10), 4) * 1e5),
+  "Numeric decimal" = abs(round(rnorm(10), 5)),
+  "This column has a very long name that means that the column width needs to be widened" = 1:10,
+  Notes = c("[note 2]", rep(NA_character_, 4), "[note 3]", rep(NA_character_, 4)),
+  check.names = FALSE
+)
+
+table_2_df <- data.frame(Category = LETTERS[1:10], Numeric = 1:10)
+
+demo_a11ytable <-
+  a11ytables::create_a11ytable(
+    tab_titles = c("Cover", "Contents", "Notes", "Table_1", "Table_2"),
+    sheet_types = c("cover", "contents", "notes", "tables", "tables"),
+    sheet_titles = c(
+      "The 'a11ytables' Demo Workbook",
+      "Table of contents",
+      "Notes",
+      "Table_1: First Example Sheet",
+      "Table_2: Second Example Sheet"
+    ),
+    blank_cells = c(
+      rep(NA_character_, 3),
+      "Blank cells indicate that there's no note in that row.",
+      NA_character_
+    ),
+    custom_rows = list(
+      NA_character_,
+      NA_character_,
+      "A custom row.",
+      c(
+        "First custom row [with a hyperlink.](https://co-analysis.github.io/a11ytables/)",
+        "Second custom row."
+      ),
+      "A custom row."
+    ),
+    sources = c(
+      rep(NA_character_, 3),
+      "[The Source Material, 2024.](https://co-analysis.github.io/a11ytables/)",
+      "The Source Material, 2024."
+    ),
+    tables = list(cover_list, contents_df, notes_df, table_1_df, table_2_df)
+  )
+
+demo_df <- as.data.frame(demo_a11ytable)
+
+demo_workbook <- generate_workbook(demo_a11ytable)
+
+# Write to data/
+usethis::use_data(demo_df, overwrite = TRUE)
+usethis::use_data(demo_a11ytable, overwrite = TRUE)
+usethis::use_data(demo_workbook, overwrite = TRUE)
+
+
+# mtcars_df and mtcars_df2 (superseded in v0.3)  --------------------------
+
 
 library(tibble)
 
@@ -80,7 +171,7 @@ mtcars_df <- tibble(
   table = list(cover_df, contents_df, notes_df, stats_df_1, stats_df_2)
 )
 
-# Using cover_list as the table input for the cover
+# Using cover_list as the table input for the cover, introduced in v0.2
 mtcars_df2 <- tibble(
   tab_title = c("Cover", "Contents", "Notes", "Table_1", "Table_2"),
   sheet_type = c("cover", "contents", "notes", "tables", "tables"),