+++ tests/testthat/test-a11ytable.R
@@ -1,16 +1,15 @@
 
-test_that("a11ytable can be created by hand (with df for cover)", {
+test_that("a11ytable can be created by hand (with list for cover)", {
 
-  # Uses mtcars_df, which has a data.frame containing cover information in the
+  # Uses demo_df, which has a list containing cover information in the
   # 'table' column.
 
   x <- suppressWarnings(
     create_a11ytable(
-      tab_titles   = mtcars_df$tab_title,
-      sheet_types  = mtcars_df$sheet_type,
-      sheet_titles = mtcars_df$sheet_title,
-      sources      = mtcars_df$source,
-      tables       = mtcars_df$table
+      tab_titles   = demo_df$tab_title,
+      sheet_types  = demo_df$sheet_type,
+      sheet_titles = demo_df$sheet_title,
+      tables       = demo_df$table
     )
   )
 
@@ -20,29 +19,27 @@ test_that("a11ytable can be created by hand (with df for cover)", {
   expect_error(
     suppressWarnings(
       create_a11ytable(
-        tab_titles   = mtcars_df$tab_title,
+        tab_titles   = demo_df$tab_title,
         sheet_types  = "x",
-        sheet_titles = mtcars_df$sheet_title,
-        sources      = mtcars_df$source,
-        tables       = mtcars_df$table
+        sheet_titles = demo_df$sheet_title,
+        tables       = demo_df$table
       )
     )
   )
 
 })
 
-test_that("a11ytable can be created by hand (with list for cover)", {
-
-  # Uses mtcars_df, which has a list containing cover information in the
-  # 'table' column.
+test_that("a11ytable can be created by hand (with df for cover)", {
 
   x <- suppressWarnings(
     create_a11ytable(
-      tab_titles   = mtcars_df2$tab_title,
-      sheet_types  = mtcars_df2$sheet_type,
-      sheet_titles = mtcars_df2$sheet_title,
-      sources      = mtcars_df2$source,
-      tables       = mtcars_df2$table
+      tab_titles   = demo_df$tab_title,
+      sheet_types  = demo_df$sheet_type,
+      sheet_titles = demo_df$sheet_title,
+      tables       = c(
+        list(data.frame("Section 1" = "Text.", "Section 2" = "Text.")),
+        demo_df$table[2:length(demo_df$table)]
+      )
     )
   )
 
@@ -52,11 +49,10 @@ test_that("a11ytable can be created by hand (with list for cover)", {
   expect_error(
     suppressWarnings(
       create_a11ytable(
-        tab_titles   = mtcars_df2$tab_title,
+        tab_titles   = demo_df$tab_title,
         sheet_types  = "x",
-        sheet_titles = mtcars_df2$sheet_title,
-        sources      = mtcars_df2$source,
-        tables       = mtcars_df2$table
+        sheet_titles = demo_df$sheet_title,
+        tables       = demo_df$table
       )
     )
   )
@@ -67,11 +63,10 @@ test_that("strings are not converted to factors", {
 
   x <- suppressWarnings(
     create_a11ytable(
-      tab_titles   = mtcars_df$tab_title,
-      sheet_types  = mtcars_df$sheet_type,
-      sheet_titles = mtcars_df$sheet_title,
-      sources      = mtcars_df$source,
-      tables       = mtcars_df$table
+      tab_titles   = demo_df$tab_title,
+      sheet_types  = demo_df$sheet_type,
+      sheet_titles = demo_df$sheet_title,
+      tables       = demo_df$table
     )
   )
 
@@ -84,7 +79,7 @@ test_that("strings are not converted to factors", {
 
 test_that("suitable objects can be coerced", {
 
-  x <- suppressWarnings(as_a11ytable(mtcars_df))
+  x <- suppressWarnings(as_a11ytable(demo_df))
 
   expect_s3_class(x, class = "a11ytable")
   expect_identical(class(x), c("a11ytable", "tbl", "data.frame"))
@@ -99,45 +94,57 @@ test_that("suitable objects can be coerced", {
 
 test_that("class validation works", {
 
-  expect_length(suppressWarnings(as_a11ytable(mtcars_df)), 6)
+  expect_length(suppressWarnings(as_a11ytable(demo_df)), 7)
 
   expect_error(as_a11ytable(1))
   expect_error(as_a11ytable("x"))
   expect_error(as_a11ytable(list()))
   expect_error(as_a11ytable(data.frame()))
 
-  x <- mtcars_df
+  x <- demo_df
   names(x)[1] <- "foo"
   expect_error(as_a11ytable(x))
 
-  x <- mtcars_df
+  x <- demo_df
   x[["table"]] <- as.character(x[["table"]])
   expect_error(as_a11ytable(x))
 
-  x <- mtcars_df[, 1:4]
+  x <- demo_df[, 1:4]
   expect_error(as_a11ytable(x))
 
-  x <- mtcars_df[1, ]
+  x <- demo_df[1, ]
   expect_error(as_a11ytable(x))
 
-  x <- mtcars_df
+  x <- demo_df
   x[x$sheet_type %in% c("cover", "contents"), "sheet_type"] <- "foo"
   expect_error(as_a11ytable(x))
 
-  x <- mtcars_df
+  x <- demo_df
   x[x$tab_title == "Table_2", "sheet_type"] <- "foo"
   expect_error(as_a11ytable(x))
 
-  x <- mtcars_df
+  x <- demo_df
   x$sheet_type <- NA_character_
   expect_error(as_a11ytable(x))
 
-  x <- mtcars_df
+  x <- demo_df
+  x$custom_rows <- NA_character_
+  expect_error(as_a11ytable(x))
+
+  x <- demo_df
+  x$custom_rows <- rep(list(1), nrow(x))
+  expect_error(as_a11ytable(x))
+
+  x <- demo_df
   x[x$tab_title == "Table_2", "tab_title"] <-
     "Lorem_ipsum_dolor_sit_amet__consectetur_adipiscing"
   expect_warning(as_a11ytable(x))
 
-  x <- mtcars_df
+  x <- demo_df
+  x[x$tab_title == "Table_2", "tab_title"] <- "!?"
+  expect_warning(as_a11ytable(x))
+
+  x <- demo_df
   x[x$sheet_type == "notes", "table"][[1]] <-
     list(
       data.frame(
@@ -152,14 +159,14 @@ test_that("class validation works", {
 
 test_that("summary method works", {
 
-  x <- suppressWarnings(as_a11ytable(mtcars_df))
+  x <- suppressWarnings(as_a11ytable(demo_df))
   expect_output(summary(x))
 
 })
 
 test_that("absence of note sheets doesn't prevent a11ytable formation", {
 
-  df <- mtcars_df[mtcars_df$sheet_type != "notes", ]
+  df <- demo_df[demo_df$sheet_type != "notes", ]
   suppressWarnings(x <- as_a11ytable(df))
 
   expect_s3_class(x, "a11ytable")
@@ -169,17 +176,15 @@ test_that("absence of note sheets doesn't prevent a11ytable formation", {
 
 test_that("tab_titles are unique", {
 
-  mtcars_df[mtcars_df$tab_title == "Table_2", "tab_title"] <- "Table_1"
+  demo_df[demo_df$tab_title == "Table_2", "tab_title"] <- "Table_1"
 
   expect_error(
     with(
-      mtcars_df,
+      demo_df,
       create_a11ytable(
         tab_titles   = tab_title,
         sheet_types  = sheet_type,
         sheet_titles = sheet_title,
-        blank_cells  = blank_cells,
-        sources      = source,
         tables       = table
       )
     )
@@ -193,7 +198,7 @@ test_that("tbl output looks as intended", {
     tab_titles = LETTERS[1:3],
     sheet_type = c("cover", "contents", "tables"),
     sheet_titles = LETTERS[1:3],
-    sources = c(rep(NA_character_, 2), "x."),
+    sources = c(NA_character_, NA_character_, "Source"),
     tables = list(
       data.frame(x = "x"),
       data.frame(tab = "x", title = "x"),
@@ -209,13 +214,13 @@ test_that("input other than data.frame is intercepted during validation", {
 
   expect_error(.validate_a11ytable("x"))
 
-  x <- mtcars_df; x[, "table"] <- "x"
+  x <- demo_df; x[, "table"] <- "x"
   expect_error(
     as_a11ytable(x),
     "Column 'table' must be a listcol of data.frame objects."
   )
 
-  y <- subset(mtcars_df, select = -table)
+  y <- subset(demo_df, select = -table)
   y[, "table"] <- list(rep(list("x"), nrow(y)))
   expect_error(
     as_a11ytable(y),
@@ -227,11 +232,11 @@ test_that("input other than data.frame is intercepted during validation", {
 test_that("only one cover, contents, notes can be used", {
 
   cover_dupe <-
-    rbind(mtcars_df, mtcars_df[mtcars_df$sheet_type == "cover", ])
+    rbind(demo_df, demo_df[demo_df$sheet_type == "cover", ])
   contents_dupe <-
-    rbind(mtcars_df, mtcars_df[mtcars_df$sheet_type == "contents", ])
+    rbind(demo_df, demo_df[demo_df$sheet_type == "contents", ])
   notes_dupe <-
-    rbind(mtcars_df, mtcars_df[mtcars_df$sheet_type == "notes", ])
+    rbind(demo_df, demo_df[demo_df$sheet_type == "notes", ])
 
   expect_error(as_a11ytable(cover_dupe))
   expect_error(as_a11ytable(contents_dupe))
@@ -241,10 +246,10 @@ test_that("only one cover, contents, notes can be used", {
 
 test_that("NAs in certain columns cause failure", {
 
-  tab_na   <- mtcars_df; tab_na$tab_title     <- NA_character_
-  type_na  <- mtcars_df; type_na$sheet_type   <- NA_character_
-  title_na <- mtcars_df; title_na$sheet_title <- NA_character_
-  table_na <- mtcars_df; table_na$table       <- NA_character_
+  tab_na   <- demo_df; tab_na$tab_title     <- NA_character_
+  type_na  <- demo_df; type_na$sheet_type   <- NA_character_
+  title_na <- demo_df; title_na$sheet_title <- NA_character_
+  table_na <- demo_df; table_na$table       <- NA_character_
 
   expect_error(as_a11ytable(tab_na))
   expect_error(as_a11ytable(type_na))
@@ -255,27 +260,18 @@ test_that("NAs in certain columns cause failure", {
 
 test_that("Note mismatch is caught", {
 
-  x <- mtcars_df[!mtcars_df$tab_title == "Table_1", ]
+  x <- demo_df[!demo_df$tab_title == "Table_1", ]
   x[x$sheet_type == "contents", "table"][[1]] <-
     list(data.frame(x = c("x", "y"), y = c("x", "y")))
 
   expect_warning(as_a11ytable(x), "You have a 'notes' sheet")
 
-  # y <- mtcars_df[!mtcars_df$tab_title == "Table_2", ]
-  # y[y$sheet_type == "contents", "table"][[1]] <-
-  #   list(data.frame(x = c("x", "y"), y = c("x", "y")))
-  # table_extra_note <- y[y$tab_title == "Table_1", "table"][[1]]
-  # names(table_extra_note)[1] <- "Car [note 3]"
-  # y[y$tab_title == "Table_1", "table"][[1]] <- list(table_extra_note)
-  #
-  # expect_warning(as_a11ytable(y), "Some notes are in the tables")
-
-  z <- mtcars_df[!mtcars_df$tab_title == "Table_2", ]
+  z <- demo_df[!demo_df$tab_title == "Table_2", ]
   z[z$sheet_type == "contents", "table"][[1]] <-
     list(data.frame(x = c("x", "y"), y = c("x", "y")))
   z[z$sheet_type == "notes", "table"][[1]] <- list(
     data.frame(
-      `Note number` = paste0("[note ", 1:3, "]"),
+      `Note number` = paste0("[note ", 1:4, "]"),
       `Note text` = "x",
       check.names = FALSE
     )
@@ -288,9 +284,9 @@ test_that("Note mismatch is caught", {
 
 test_that("warning is raised if a source statement is missing", {
 
-  mtcars_df[mtcars_df$tab_title == "Table_1", "source"] <- NA_character_
+  demo_df[demo_df$tab_title == "Table_1", "source"] <- NA_character_
   expect_warning(
-    as_a11ytable(mtcars_df),
+    as_a11ytable(demo_df),
     "One of your tables is missing a source statement."
   )
 
@@ -298,9 +294,9 @@ test_that("warning is raised if a source statement is missing", {
 
 test_that("warning is raised if there's no blank cells but there is a reason", {
 
-  mtcars_df[mtcars_df$tab_title == "Table_2", "blank_cells"] <- "x"
+  demo_df[demo_df$tab_title == "Table_2", "blank_cells"] <- "x"
   expect_warning(
-    as_a11ytable(mtcars_df),
+    as_a11ytable(demo_df),
     "There's no blank cells in these tables"
   )
 
@@ -329,11 +325,11 @@ test_that("tab titles are cleaned and warnings provided", {
   expect_warning(.clean_tab_titles("Table!@£#$%^&*(){}[]-=+;:'\"\\|<>,.~`/?1"))
   expect_warning(.clean_tab_titles(long_title))
 
-  x <- mtcars_df
+  x <- demo_df
   x[1, "tab_title"] <- long_title
   expect_warning(as_a11ytable(x))
 
-  y <- mtcars_df
+  y <- demo_df
   y[1, "tab_title"] <- "Cover!"
   expect_warning(as_a11ytable(y))
 
@@ -341,14 +337,14 @@ test_that("tab titles are cleaned and warnings provided", {
 
 test_that("input column names are okay", {
 
-  names(mtcars_df)[1] <- "x"
-  expect_error(as_a11ytable(mtcars_df))
+  names(demo_df)[1] <- "x"
+  expect_error(as_a11ytable(demo_df))
 
 })
 
 test_that("character class columns are caught if not character class", {
 
-  mtcars_df[, "sheet_type"] <- 1:nrow(mtcars_df)
-  expect_error(as_a11ytable(mtcars_df))
+  demo_df[, "sheet_type"] <- 1:nrow(demo_df)
+  expect_error(as_a11ytable(demo_df))
 
 })